requirements.txt
absl-py==1.3.0
continuum==1.2.4
filelock==3.8.0
lightning==1.9.4
numpy==1.23.4
omegaconf==2.2.3
tensorboard==2.10.1
timm==0.5.4
torch==1.12.1
torchinfo==1.7.1
torchmetrics==0.10.2
torchvision==0.13.1
typing_extensions>=4.0
